roachtest: move HealthChecker, DiskUsageTracker/Logger to roachtestutil pkg
Epic: none

Release note: none
