+++ pkg/cmd/roachtest/roachtestutil/BUILD.bazel
@@ -5,6 +5,8 @@ go_library(
     name = "roachtestutil",
     srcs = [
         "commandbuilder.go",
+        "disk_usage.go",
+        "health_checker.go",
         "jaeger.go",
         "validation_check.go",
     ],
@@ -12,8 +14,13 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/cmd/roachtest/cluster",
+        "//pkg/cmd/roachtest/option",
+        "//pkg/cmd/roachtest/test",
+        "//pkg/roachprod/install",
         "//pkg/roachprod/logger",
         "//pkg/testutils/sqlutils",
+        "//pkg/util/humanizeutil",
+        "//pkg/util/timeutil",
         "@com_github_cockroachdb_errors//:errors",
     ],
 )
