+++ pkg/cmd/roachtest/roachtestutil/health_checker.go
@@ -1,0 +1,141 @@
+// Copyright 2023 The Cockroach Authors.
+//
+// Use of this software is governed by the Business Source License
+// included in the file licenses/BSL.txt.
+//
+// As of the Change Date specified in that file, in accordance with
+// the Business Source License, use of this software will be governed
+// by the Apache License, Version 2.0, included in the file
+// licenses/APL.txt.
+
+package roachtestutil
+
+import (
+	"bytes"
+	"context"
+	gosql "database/sql"
+	"fmt"
+	"math/rand"
+	"text/tabwriter"
+	"time"
+
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/cluster"
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/option"
+	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/test"
+	"github.com/cockroachdb/cockroach/pkg/util/timeutil"
+	"github.com/cockroachdb/errors"
+)
+
+// HealthChecker runs a regular check that verifies that a specified subset
+// of (CockroachDB) nodes look "very healthy". That is, there are no stuck
+// proposals, liveness problems, or whatever else might get added in the
+// future.
+type HealthChecker struct {
+	t      test.Test
+	c      cluster.Cluster
+	nodes  option.NodeListOption
+	doneCh chan struct{}
+}
+
+// NewHealthChecker returns a populated HealthChecker.
+func NewHealthChecker(t test.Test, c cluster.Cluster, nodes option.NodeListOption) *HealthChecker {
+	return &HealthChecker{
+		t:      t,
+		c:      c,
+		nodes:  nodes,
+		doneCh: make(chan struct{}),
+	}
+}
+
+// Done signals the HealthChecker's Runner to shut down.
+func (hc *HealthChecker) Done() {
+	close(hc.doneCh)
+}
+
+type gossipAlert struct {
+	NodeID, StoreID       int
+	Category, Description string
+	Value                 float64
+}
+
+type gossipAlerts []gossipAlert
+
+func (g gossipAlerts) String() string {
+	var buf bytes.Buffer
+	tw := tabwriter.NewWriter(&buf, 2, 1, 2, ' ', 0)
+
+	for _, a := range g {
+		fmt.Fprintf(tw, "n%d/s%d\t%.2f\t%s\t%s\n", a.NodeID, a.StoreID, a.Value, a.Category, a.Description)
+	}
+	_ = tw.Flush()
+	return buf.String()
+}
+
+// Runner makes sure the gossip_alerts table is empty at all times.
+//
+// TODO(tschottdorf): actually let this fail the test instead of logging complaints.
+func (hc *HealthChecker) Runner(ctx context.Context) (err error) {
+	logger, err := hc.t.L().ChildLogger("health")
+	if err != nil {
+		return err
+	}
+	defer func() {
+		logger.Printf("health check terminated with %v\n", err)
+	}()
+
+	ticker := time.NewTicker(5 * time.Second)
+	defer ticker.Stop()
+
+	for {
+		select {
+		case <-ctx.Done():
+			return ctx.Err()
+		case <-hc.doneCh:
+			return nil
+		case <-ticker.C:
+		}
+
+		tBegin := timeutil.Now()
+
+		nodeIdx := 1 + rand.Intn(len(hc.nodes))
+		db, err := hc.c.ConnE(ctx, hc.t.L(), nodeIdx)
+		if err != nil {
+			return err
+		}
+		// TODO(tschottdorf): remove replicate queue failures when the cluster first starts.
+		// Ditto queue.raftsnapshot.process.failure.
+		_, err = db.Exec(`USE system`)
+		if err != nil {
+			return err
+		}
+		rows, err := db.QueryContext(ctx, `SELECT * FROM crdb_internal.gossip_alerts ORDER BY node_id ASC, store_id ASC `)
+		_ = db.Close()
+		if err != nil {
+			return err
+		}
+		var rr gossipAlerts
+		for rows.Next() {
+			a := gossipAlert{StoreID: -1}
+			var storeID gosql.NullInt64
+			if err := rows.Scan(&a.NodeID, &storeID, &a.Category, &a.Description, &a.Value); err != nil {
+				return err
+			}
+			if storeID.Valid {
+				a.StoreID = int(storeID.Int64)
+			}
+			rr = append(rr, a)
+		}
+		if len(rr) > 0 {
+			logger.Printf(rr.String() + "\n")
+			// TODO(tschottdorf): see method comment.
+			// return errors.New(rr.String())
+		}
+
+		if elapsed := timeutil.Since(tBegin); elapsed > 10*time.Second {
+			err := errors.Errorf("health check against node %d took %s", nodeIdx, elapsed)
+			logger.Printf("%+v", err)
+			// TODO(tschottdorf): see method comment.
+			// return err
+		}
+	}
+}
