+++ pkg/cmd/roachtest/tests/util_disk_usage.go
@@ -9,62 +9,3 @@
 // licenses/APL.txt.
 
 package tests
-
-import (
-	"context"
-	"strconv"
-	"strings"
-
-	"github.com/cockroachdb/cockroach/pkg/cmd/roachtest/cluster"
-	"github.com/cockroachdb/cockroach/pkg/roachprod/install"
-	"github.com/cockroachdb/cockroach/pkg/roachprod/logger"
-)
-
-// getDiskUsageInBytes does what's on the tin. nodeIdx starts at one.
-func getDiskUsageInBytes(
-	ctx context.Context, c cluster.Cluster, logger *logger.Logger, nodeIdx int,
-) (int, error) {
-	var result install.RunResultDetails
-	for {
-		var err error
-		// `du` can warn if files get removed out from under it (which
-		// happens during RocksDB compactions, for example). Discard its
-		// stderr to avoid breaking Atoi later.
-		// TODO(bdarnell): Refactor this stack to not combine stdout and
-		// stderr so we don't need to do this (and the Warning check
-		// below).
-		result, err = c.RunWithDetailsSingleNode(
-			ctx,
-			logger,
-			c.Node(nodeIdx),
-			"du -sk {store-dir} 2>/dev/null | grep -oE '^[0-9]+'",
-		)
-		if err != nil {
-			if ctx.Err() != nil {
-				return 0, ctx.Err()
-			}
-			// If `du` fails, retry.
-			// TODO(bdarnell): is this worth doing? It was originally added
-			// because of the "files removed out from under it" problem, but
-			// that doesn't result in a command failure, just a stderr
-			// message.
-			logger.Printf("retrying disk usage computation after spurious error: %s", err)
-			continue
-		}
-
-		break
-	}
-
-	// We need this check because sometimes the first line of the roachprod output is a warning
-	// about adding an ip to a list of known hosts.
-	if strings.Contains(result.Stdout, "Warning") {
-		result.Stdout = strings.Split(result.Stdout, "\n")[1]
-	}
-
-	size, err := strconv.Atoi(strings.TrimSpace(result.Stdout))
-	if err != nil {
-		return 0, err
-	}
-
-	return size * 1024, nil
-}
